Move animation drawable to parent class
