+++ android/java/org/chromium/chrome/browser/crypto_wallet/fragments/onboarding/BaseOnboardingWalletFragment.java
@@ -5,18 +5,14 @@
 
 package org.chromium.chrome.browser.crypto_wallet.fragments.onboarding;
 
-import android.graphics.drawable.AnimationDrawable;
 import android.os.Bundle;
-import android.view.View;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.widget.AppCompatButton;
-import androidx.core.content.ContextCompat;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.lifecycle.ViewModelStoreOwner;
 
-import org.chromium.chrome.R;
 import org.chromium.chrome.browser.crypto_wallet.fragments.BaseWalletNextPageFragment;
 import org.chromium.chrome.browser.crypto_wallet.model.OnboardingViewModel;
 
@@ -40,8 +36,6 @@ public abstract class BaseOnboardingWalletFragment extends BaseWalletNextPageFra
         return true;
     }
 
-    @Nullable private AnimationDrawable mAnimationDrawable;
-
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -58,29 +52,6 @@ public abstract class BaseOnboardingWalletFragment extends BaseWalletNextPageFra
             mOnNextPage.showCloseButton(canBeClosed());
             // Show or hide back icon depending on the fragment configuration.
             mOnNextPage.showBackButton(canNavigateBack());
-        }
-        if (mAnimationDrawable != null) {
-            mAnimationDrawable.start();
-        }
-    }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-        if (mAnimationDrawable != null) {
-            mAnimationDrawable.stop();
-        }
-    }
-
-    protected void setAnimatedBackground(@NonNull final View rootView) {
-        mAnimationDrawable =
-                (AnimationDrawable)
-                        ContextCompat.getDrawable(
-                                requireContext(), R.drawable.onboarding_gradient_animation);
-        if (mAnimationDrawable != null) {
-            rootView.setBackground(mAnimationDrawable);
-            mAnimationDrawable.setEnterFadeDuration(10);
-            mAnimationDrawable.setExitFadeDuration(5000);
         }
     }
